+++ fs/walk.ts
@@ -3,45 +3,8 @@
 // https://golang.org/pkg/path/filepath/#Walk
 // Copyright 2009 The Go Authors. All rights reserved. BSD license.
 import { assert } from "../_util/assert.ts";
-import { basename, join, normalize } from "../path/mod.ts";
-
-/** Create WalkEntry for the `path` synchronously */
-export function _createWalkEntrySync(path: string): WalkEntry {
-  path = normalize(path);
-  const name = basename(path);
-  const info = Deno.statSync(path);
-  return {
-    path,
-    name,
-    isFile: info.isFile,
-    isDirectory: info.isDirectory,
-    isSymlink: info.isSymlink,
-  };
-}
-
-/** Create WalkEntry for the `path` asynchronously */
-export async function _createWalkEntry(path: string): Promise<WalkEntry> {
-  path = normalize(path);
-  const name = basename(path);
-  const info = await Deno.stat(path);
-  return {
-    path,
-    name,
-    isFile: info.isFile,
-    isDirectory: info.isDirectory,
-    isSymlink: info.isSymlink,
-  };
-}
-
-export interface WalkOptions {
-  maxDepth?: number;
-  includeFiles?: boolean;
-  includeDirs?: boolean;
-  followSymlinks?: boolean;
-  exts?: string[];
-  match?: RegExp[];
-  skip?: RegExp[];
-}
+import { join, normalize } from "../path/mod.ts";
+import { createWalkEntry, createWalkEntrySync, WalkEntry } from "./_util.ts";
 
 function include(
   path: string,
@@ -73,9 +36,16 @@ function wrapErrorWithRootPath(err: unknown, root: string) {
   return e;
 }
 
-export interface WalkEntry extends Deno.DirEntry {
-  path: string;
+export interface WalkOptions {
+  maxDepth?: number;
+  includeFiles?: boolean;
+  includeDirs?: boolean;
+  followSymlinks?: boolean;
+  exts?: string[];
+  match?: RegExp[];
+  skip?: RegExp[];
 }
+export type { WalkEntry };
 
 /** Walks the file tree rooted at root, yielding each file or directory in the
  * tree filtered according to the given options. The files are walked in lexical
@@ -117,7 +87,7 @@ export async function* walk(
     return;
   }
   if (includeDirs && include(root, exts, match, skip)) {
-    yield await _createWalkEntry(root);
+    yield await createWalkEntry(root);
   }
   if (maxDepth < 1 || !include(root, undefined, undefined, skip)) {
     return;
@@ -174,7 +144,7 @@ export function* walkSync(
     return;
   }
   if (includeDirs && include(root, exts, match, skip)) {
-    yield _createWalkEntrySync(root);
+    yield createWalkEntrySync(root);
   }
   if (maxDepth < 1 || !include(root, undefined, undefined, skip)) {
     return;
